+++ lua/custom/utils.lua
@@ -1,0 +1,5 @@
+local Utils = {}
+
+Utils.isWindows = vim.fn.has 'win32' == 1
+
+return Utils