+++ lua/custom/config/lsp.lua
@@ -135,8 +135,6 @@ vim.api.nvim_create_autocmd('LspAttach', {
 vim.diagnostic.config {
   severity_sort = true,
   float = { border = 'rounded', source = 'if_many' },
-  -- [[deprecated]] <https://github.com/nvim-lua/kickstart.nvim/pull/1550>
-  -- underline = { severity = vim.diagnostic.severity.ERROR },
   signs = vim.g.have_nerd_font and {
     text = {
       [vim.diagnostic.severity.ERROR] = '󰅚 ',
@@ -148,16 +146,6 @@ vim.diagnostic.config {
   virtual_text = {
     source = 'if_many',
     spacing = 2,
-    -- [[deprecated]] <https://github.com/nvim-lua/kickstart.nvim/pull/1550>
-    -- format = function(diagnostic)
-    --   local diagnostic_message = {
-    --     [vim.diagnostic.severity.ERROR] = diagnostic.message,
-    --     [vim.diagnostic.severity.WARN] = diagnostic.message,
-    --     [vim.diagnostic.severity.INFO] = diagnostic.message,
-    --     [vim.diagnostic.severity.HINT] = diagnostic.message,
-    --   }
-    --   return diagnostic_message[diagnostic.severity]
-    -- end,
   },
 }