+++ lua/configs/git.lua
@@ -1,0 +1,16 @@
+vim.pack.add {
+  'https://github.com/tpope/vim-fugitive',
+  'https://github.com/tpope/vim-rhubarb',
+}
+
+vim.pack.add { 'https://github.com/lewis6991/gitsigns.nvim' }
+
+require('gitsigns').setup {
+  signs = {
+    add = { text = '+' },
+    change = { text = '~' },
+    delete = { text = '_' },
+    topdelete = { text = '‾' },
+    changedelete = { text = '~' },
+  },
+}