+++ lua/configs/init.lua
@@ -1,0 +1,17 @@
+require 'configs.options' -- 原生配置。
+require 'configs.commands' -- 自定义命令。
+require 'configs.icon' -- 图标。
+require 'configs.colorscheme' -- 主题。
+require 'configs.statusline' -- 底部状态栏。
+require 'configs.treesitter' -- 语法树高亮。
+require 'configs.filetree' -- 目录系统。
+require 'configs.git' -- Git。
+require 'configs.indent' -- 缩进。
+require 'configs.tools' -- 工具。
+require 'configs.search' -- 搜索。
+require 'configs.lsp' -- 语言服务协议。
+require 'configs.autoformat' -- 自动格式化。
+require 'configs.completion' -- 代码补全。
+require 'configs.lint' -- 语法检查。
+require 'configs.preview' -- 预览。
+require 'configs.tabline' -- 顶部 buffer 栏（和 vim.pack 窗口冲突所以放到最后）。