+++ lua/custom/config/init.lua
@@ -1,6 +1,0 @@
-require 'custom.config.highlight'
-
-require 'custom.config.treesitter'
-require 'custom.config.indent'
-require 'custom.config.lsp'
-require 'custom.config.markdown'